refactor: 입력 처리 리팩토링 및 패키지 재구성
- InputView 클래스 생성: 자동차 이름과 시도 횟수에 대한 사용자 입력을 처리하여 관심사의 분리를 개선
- Car 클래스를 public으로 변경하여 더 넓은 접근을 허용
- InputValidator 클래스와 OutputView 클래스를 더 적절한 패키지로 이름 변경 및 이동
