+++ src/main/java/racingcar/controller/GameController.java
@@ -2,9 +2,15 @@ package racingcar.controller;
 
 import racingcar.domain.Car;
 import racingcar.domain.RacingGame;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class GameController {
+
+
+
+
 }
